Изм 6. разделение классов фрагмента home и ratings - зачистка кода.
