+++ app/src/main/java/ru/geekbrains/androidonkotlin/hw/mymovie/ui/home/HomeBasicAdapter.kt
@@ -3,48 +3,40 @@ package ru.geekbrains.androidonkotlin.hw.mymovie.ui.home
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.ViewModelProvider
+import androidx.fragment.app.FragmentManager
 import androidx.recyclerview.widget.RecyclerView
 import ru.geekbrains.androidonkotlin.hw.mymovie.R
-import ru.geekbrains.androidonkotlin.hw.mymovie.domain.MovieTmdb
 import ru.geekbrains.androidonkotlin.hw.mymovie.ui.GroupResponseObject
-import ru.geekbrains.androidonkotlin.hw.mymovie.ui.interfaces.OnLoadMoreMovies
 
 class HomeBasicAdapter(_fragment: Fragment) :
     RecyclerView.Adapter<HomeBasicViewHolder>() {
-    private val homeViewModel: HomeViewModel =
-        ViewModelProvider(_fragment).get(HomeViewModel::class.java)
     val fragment: Fragment = _fragment
     var items: ArrayList<GroupResponseObject> = ArrayList()
+    private val fragmentManager: FragmentManager = fragment.childFragmentManager
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): HomeBasicViewHolder {
-        val root =
-            LayoutInflater.from(parent.context).inflate(R.layout.basic_item_home, parent, false)
+        val root = LayoutInflater.from(parent.context).inflate(R.layout.basic_item_home, parent, false)
         return HomeBasicViewHolder(root)
     }
 
     override fun onBindViewHolder(holder: HomeBasicViewHolder, position: Int) {
-        val item = items.get(position)
+        val item = items[position]
         holder.basicTitle.text = item.nameGroupResponse
-        //работа с вложенным адаптером
-        val currentRO = homeViewModel.arrGroupList[position]
-        val currentLiveData: LiveData<List<MovieTmdb>> = currentRO.currentLiveData
-        currentLiveData.observe(fragment.viewLifecycleOwner, {
-            holder.adapter.items = it
-            holder.adapter.setOnLoadMoreMoviesListener(object : OnLoadMoreMovies {
-                override fun onLoadMore() {
-                    if (currentRO.lastAnswer.page < currentRO.lastAnswer.totalPages) {
-                        currentRO.funcFetch.invoke(
-                            currentRO.standardList.toString(),
-                            currentRO.lastAnswer.page + 1, currentRO
-                        )
-                    }
-                }
-            })
-            holder.adapter.notifyDataSetChanged()
-        })
+        holder.flContainer.id = position + 100000
     }
 
     override fun getItemCount(): Int = items.size
+
+    override fun onViewAttachedToWindow(holder: HomeBasicViewHolder) {
+        attachFragmentToContainer(holder.flContainer.id, holder.adapterPosition)
+        super.onViewAttachedToWindow(holder)
+    }
+
+    private fun attachFragmentToContainer(containerId: Int, fragmentListId: Int) {
+        if (fragmentManager.findFragmentByTag(containerId.toString()) == null) {
+            val fragment = HomeInnerFragment(items[fragmentListId])
+            fragmentManager.beginTransaction().add(containerId, fragment, containerId.toString())
+                .commitNowAllowingStateLoss()
+        }
+    }
 }
