+++ app/src/main/java/ru/geekbrains/androidonkotlin/hw/mymovie/ui/home/HomeBasicViewHolder.kt
@@ -1,6 +1,7 @@
 package ru.geekbrains.androidonkotlin.hw.mymovie.ui.home
 
 import android.view.View
+import android.widget.FrameLayout
 import android.widget.TextView
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
@@ -8,13 +9,5 @@ import ru.geekbrains.androidonkotlin.hw.mymovie.R
 
 class HomeBasicViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
     val basicTitle: TextView = itemView.findViewById(R.id.basicTitle)
-    var adapter: HomeInnerAdapter = HomeInnerAdapter()
-
-    init {
-        val homeInnerRecyclerView: RecyclerView =
-            itemView.findViewById(R.id.home_inner_list)
-        homeInnerRecyclerView.adapter = adapter
-        homeInnerRecyclerView.layoutManager =
-            LinearLayoutManager(itemView.context, LinearLayoutManager.HORIZONTAL, false)
-    }
+    val flContainer: FrameLayout = itemView.findViewById(R.id.flContainer)
 }
