+++ app/src/main/java/ru/geekbrains/androidonkotlin/hw/mymovie/ui/ratings/RatingBasicViewHolder.kt
@@ -11,10 +11,8 @@ class RatingBasicViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) 
     var adapter: RatingInnerAdapter = RatingInnerAdapter()
 
     init {
-        val homeInnerRecyclerView: RecyclerView =
-            itemView.findViewById(R.id.home_inner_list)
-        homeInnerRecyclerView.adapter = adapter
-        homeInnerRecyclerView.layoutManager =
-            LinearLayoutManager(itemView.context, LinearLayoutManager.HORIZONTAL, false)
+        val ratingsInnerRecyclerView: RecyclerView = itemView.findViewById(R.id.home_inner_list)
+        ratingsInnerRecyclerView.adapter = adapter
+        ratingsInnerRecyclerView.layoutManager = LinearLayoutManager(itemView.context, LinearLayoutManager.HORIZONTAL, false)
     }
 }
