+++ app/src/main/java/ru/geekbrains/androidonkotlin/hw/mymovie/ui/ratings/RatingsFragment.kt
@@ -48,8 +48,8 @@ class RatingsFragment : Fragment() {
         })
     }
 
-    override fun onPause() {
-        super.onPause()
+    override fun onStop() {
+        super.onStop()
         ratingsViewModel.ratingBasicStructureLiveData.removeObservers(this)
     }
