+++ app/src/main/res/layout/basic_item_ratings.xml
@@ -10,14 +10,14 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="16dp"
-        android:text="Наименование рубрики (временно)"
+        android:text="@string/temp_group_movie_name"
         android:textSize="18sp"
         android:textStyle="bold"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
-    <FrameLayout
-        android:id="@+id/flContainer"
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/ratings_inner_list"
         android:layout_width="match_parent"
         android:layout_height="0dp"
         android:layout_marginStart="4dp"
