+++ app/src/main/res/values/strings.xml
@@ -34,4 +34,5 @@
     <string name="search_fragment_label_result">Результаты поиска...</string>
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="temp_group_movie_name">Наименование рубрики (временно)</string>
 </resources>
